+++ tasks/orchestrator/lib/erb.rb
@@ -3,10 +3,10 @@ module Orchestrator
     require 'erb'
 
     # Create Rendered File from ERB Template Locally
-    def self.create_file(file, template, acl)
+    def self.create_file(file, acl, template, variables)
       # file = full path and file name to create
       # template = full path to .erb file
-      rendered_content = Orchestrator::Erb.render_template(template)
+      rendered_content = Orchestrator::Erb.render_template(template, variables)
       Orchestrator::ConsoleOutputs.debug_message_item('[Orchestrator::Erb.create_file] rendered_content', rendered_content)
       Orchestrator::ConsoleOutputs.debug_message_item('[Orchestrator::Erb.create_file] rendered_content.class', rendered_content.class)
 
@@ -18,9 +18,9 @@ module Orchestrator
     end
 
     # Return rendered Ruby file Template object. Used for creating remote files using rendered content in heredoc shell call
-    def self.render_template(template)
+    def self.render_template(template, variables)
       # template = full path to .erb file
-      rendered_template = ERB.new(File.read(template)).result
+      rendered_template = ERB.new(File.read(template)).instance_variable_set(variables).result
       Orchestrator::ConsoleOutputs.debug_message_item('[Orchestrator::Erb.render_template] rendered_template', rendered_template)
       rendered_template
     end